feat(contacts): restructure to layout+index, add inline editing, fix navigation
- contacts.tsx → pure layout with <Outlet /> so $contactId renders standalone
- contacts/index.tsx → list page (tabs, toolbar, DataTableRoot)
- contacts-columns: add isEditable + onSave per field (name, type, email, phone)
- contacts-list: wire onUpdate mutation for inline cell editing
- Fix navigate to $contactId using useNavigate() instead of routeApi.useNavigate()
